server/requirements.txt
flask
cors
flask_cors
openai
langchain
wolframalpha
axios
python-dotenv